requirements.txt
requests
beautifulsoup4
mysql-connector-python
pandas
streamlit